+++ platform-api/src/main/java/com/platform/entity/CollectVo.java
@@ -13,7 +13,7 @@ public class CollectVo implements Serializable {
     private static final long serialVersionUID = 1L;
 
     //主键
-    private Integer id;
+    private Long id;
     //用户Id
     private Long user_id;
     //产品Id
@@ -34,7 +34,7 @@ public class CollectVo implements Serializable {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
