+++ platform-shop/src/main/java/com/platform/entity/CategoryEntity.java
@@ -11,7 +11,7 @@ package com.platform.entity;
 public class CategoryEntity extends Tree<CategoryEntity> {
 
     //主键
-    private Integer id;
+    private Long id;
     //分类名称
     private String name;
     //关键字
@@ -55,7 +55,7 @@ public class CategoryEntity extends Tree<CategoryEntity> {
     /**
      * 设置：主键
      */
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
